refactor: Box 大 variant 优化 + Copy derive 补全 + 文档重组
- Box 大 variant: ContentBlock/StreamEvent/AgentEvent/SessionEntry 减少栈内存占用
- 补全 6 个简单枚举的 Copy derive（Role/DeltaType/ToolCallStatus/TaskStatus/ErrorCategory/ToolCallRenderStatus）
- 文档重组: docs/technologies/ → docs/pi-technologies/ + 新建 docs/uncode-technologies/ 7 篇架构文档
- 新增 NOW.png 图标、bun.lock、tsconfig.json 等 Platform 相关文件

+++ crates/uncode-agent/src/tools/read.rs
@@ -81,7 +81,7 @@ impl ToolExecutor for ReadTool {
         }
 
         // Check file size BEFORE reading into memory (#191)
-        if meta.len() as usize > self.max_size {
+        if meta.len() > self.max_size as u64 {
             return Err(uncode_core::error::UncodeError::Tool(format!(
                 "file too large ({} bytes, max {})",
                 meta.len(),
